1、调整用户编辑、添加、删除接口 2、调整http 419 unknown status 错误，增加ajax X-CSRF-TOKEN即可 3、删除测试json数据

+++ app/Http/Controllers/UserController.php
@@ -8,6 +8,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\User;
+use Log;
 
 class UserController extends Controller
 {
@@ -48,7 +49,11 @@ class UserController extends Controller
     public function delete($user_id)
     {
         $model = User::where(["_id"=>$user_id])->first();
-        $result = $model->delete();
+		if(!empty($model)){
+			$result = $model->delete();
+		}else{
+			$result = "";
+		}
         return $this->getBoolResult($result);
     }
